Implemented perspective correct interpolation for wireframe attributes. Still need to iron out some inconsistencies with regards to axis and vertex ordering. Once those are resolved, this can be moved over to the flat shaded code path.

+++ ZColor.cpp
@@ -3,8 +3,8 @@
 namespace ZSharp {
 void ZColor::FloatToRGB(const float R, const float G, const float B) {
   mColor = 0xFF000000;
-  mColor |= (0x00FF0000 & (static_cast<uint8>((uint32)R * (uint32)0xFFU) << 16));
-  mColor |= (0x0000FF00 & (static_cast<uint8>((uint32)G * (uint32)0xFFU) << 8));
-  mColor |= (0x000000FF & static_cast<uint8>((uint32)B * (uint32)0xFFU));
+  mColor |= (0x00FF0000 & (static_cast<uint8>(R * (float)0xFFU) << 16));
+  mColor |= (0x0000FF00 & (static_cast<uint8>(G * (float)0xFFU) << 8));
+  mColor |= (0x000000FF & static_cast<uint8>(B * (float)0xFFU));
 }
 }
